test(providers): consolidate phase-suffixed test files into per-provider files
Fold the remediation-era _pN and _phaseN test files back into each provider's
standard <provider>_test.go, removing the phase-suffix naming. Test bodies are
unchanged; topic-split test files (messages, streaming, params, etc.) are kept.
