+++ providers/ollama/ollama_test.go
@@ -380,3 +380,10 @@ func TestOllamaProvider_Embed_RejectsNonFloatEncoding(t *testing.T) {
 		t.Fatal("Embed() error = nil, want unsupported encoding_format error")
 	}
 }
+
+// TestNewOllama_RejectsInvalidBaseURL locks in base-URL validation.
+func TestNewOllama_RejectsInvalidBaseURL(t *testing.T) {
+	if _, err := New("://nope", nil); err == nil {
+		t.Fatal("New accepted an invalid base URL")
+	}
+}
