--- providers/openrouter/openrouter_p5b_test.go
@@ -1,10 +1,0 @@
-package openrouter
-
-import "testing"
-
-// TestNewOpenRouter_RejectsInvalidBaseURL locks in the base-URL validation.
-func TestNewOpenRouter_RejectsInvalidBaseURL(t *testing.T) {
-	if _, err := New("k", "://bad"); err == nil {
-		t.Fatal("New accepted an invalid base URL")
-	}
-}
